refactor: resolve python:S3752 Security Hotspot and eliminate code duplications in auth pages
